fix(nextly): count invalidated rows per dialect, and delete the unreachable RLS branch
Two findings, and the first is the SECOND time this shape has appeared: a
correct, documented, dialect-aware helper existed and was private to one
file while its neighbour got the answer wrong.

`invalidateByUser`, `invalidateByRole` and `cleanupExpired` each read
`result.rowCount`. better-sqlite3 does not set it — it reports `changes` —
so all three returned 0 on SQLite however many rows they tombstoned. The
invalidation itself was correct; only the count lied, and `?? 0` turned a
missing field into a plausible number rather than an error, so a caller
branching on it took the wrong path with nothing to debug.

`affectedRowCount` already handled all three drivers, in auth-service.ts,
exported for its own unit test and used nowhere else. It moves to
shared/lib and the three sites use it. Measured: invalidating two cached
rows now returns 2 rather than 0. The tests asserted only the observable
effect and explained in a comment why they could not assert the count;
they now assert both, because the effect alone passes when the count lies
and the count alone passes when nothing was invalidated.

Separately: the RLS validator's `CallExpression` visitor carried a branch
below its allowlist check that tried to separate global calls from member
calls, telling the author to "use member methods instead (e.g.
array.includes())". `CallExpression` is not in the allowlist, so the check
always fired first and that branch was unreachable — the message
advertised a capability that does not exist and sent anyone hitting the
real error toward a second failure. Deleted rather than enabled: blocking
every call is the ruled behaviour, and enabling it would widen an
access-control evaluator on the strength of a stale comment.

Closes finding:permission-cache-counts-zero-on-sqlite and
finding:rls-expressions-cannot-call-any-method.
